+++ CHANGELOG.md
@@ -1,3 +1,12 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+- **DOCS**: Initial Release.
+
+## 0.0.2
+
+- **DOCS**: Updated README and Example App.
+
+## 0.0.3
+
+- **DOCS**: Updated Package Licence
+